feat(blind-sig): add support for verify signature with specified public key
